Style: Make auth a layout
