+++ collection/auth.ts
@@ -27,6 +27,11 @@ export const providerMap = providers.map((provider) => {
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
     providers,
+    callbacks: {
+        signIn({ profile }) {
+            return profile?.email?.includes("docsoc") ?? false;
+        },
+    },
     pages: {
         signIn: "/login",
     },
